+++ projects/ngx-uploader-directive/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Fixed on error of sending request for uploading file http error response was not having the request id in return.
+- Fixed if request id will be passed null it was not giving error event.
+
 ## [1.1.3] - 2020-05-03
 
 ### Changed
@@ -33,7 +39,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added File select directive
 - Uploading files in a single request
 
-[1.1.3]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.1.3...1.1.2
+[Unreleased]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.1.3...HEAD
+[1.1.3]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.1.2...1.1.3
 [1.1.2]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.1.0...1.1.2
 [1.1.0]: https://github.com/jayprajapati857/ngx-uploader-directive/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/jayprajapati857/ngx-uploader-directive/releases/tag/1.0.0